make Object more useful
delegate through Deref, since this covers the typical use case of containing a smart pointer of a trait object.

this means that it's no longer good enough for the inner type to directly implement the trait. it needs to deref to something that implements the trait. but this is no big deal because:
- it's not even possible for Object to directly contain trait objects, so this is an unlikely use case
- it's easier to go from direct implementation to Deref (just Box it), than to go from Deref to direct implementation